+++ rust/main/hyperlane-core/src/types/native_token.rs
@@ -3,6 +3,8 @@
 pub struct NativeToken {
     /// The number of decimal places in token which can be expressed by denomination
     pub decimals: u32,
-    /// Denomination of the token
+    /// Token symbol (e.g. "ETH", "AVAX")
+    pub symbol: String,
+    /// Denomination of the token (used primarily by Cosmos chains)
     pub denom: String,
 }