+++ typescript/aleo-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hyperlane-xyz/aleo-sdk
 
+## 25.3.2
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@25.3.2
+- @hyperlane-xyz/provider-sdk@1.3.6
+
 ## 25.3.1
 
 ### Patch Changes