+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/helloworld
 
+## 25.3.2
+
+### Patch Changes
+
+- Updated dependencies [521d42b]
+  - @hyperlane-xyz/core@10.2.0
+  - @hyperlane-xyz/sdk@25.3.2
+  - @hyperlane-xyz/utils@25.3.2
+
 ## 25.3.1
 
 ### Patch Changes